+++ lib/src/extended/overlay/toast.dart
@@ -8,41 +8,87 @@ class Toast extends StatelessWidget {
   const Toast(
     this.message, {
     super.key,
+    this.iconStyle,
+    this.textStyle,
+    this.duration,
+    this.animationDuration,
+    this.animationStyle,
+    this.decoration,
+    this.onToastTap,
+    this.color = Colors.white,
+    this.direction = Axis.horizontal,
+    this.builder,
     this.options,
-    this.style,
-    this.icon,
   });
 
   /// 文字
   final String message;
 
-  /// UI配置
-  final ToastOptions? options;
+  /// 动画样式
+  final FlAnimationStyle? animationStyle;
+
+  /// 动画执行时间
+  final Duration? animationDuration;
+
+  /// 显示文字样式
+  final TextStyle? textStyle;
+
+  /// Toast显示时间
+  final Duration? duration;
+
+  /// Toast 装饰器 会替换 [backgroundColor]
+  final BoxDecoration? decoration;
+
+  /// Toast onTap
+  final GestureTapCallback? onToastTap;
+
+  /// 字体颜色和icon颜色
+  final Color color;
+
+  /// Toast icon direction
+  final Axis direction;
+
+  /// 重新 build  text icon
+  final ToastOptionsBuilder? builder;
 
   /// icon 样式 不传 仅显示文字
-  final ToastStyle? style;
+  final ToastIconStyle? iconStyle;
 
-  /// 自定义icon [icon] 优先于 [ToastStyle]
-  final IconData? icon;
+  /// ToastOptions
+  final ToastOptions? options;
 
   @override
   Widget build(BuildContext context) {
-    final currentOptions = FlExtended().toastOptions.merge(options);
-    Widget current =
-        currentOptions.buildText?.call(message, currentOptions.iconColor) ??
-            BText(message,
-                textAlign: TextAlign.center,
-                color: currentOptions.iconColor,
-                maxLines: 5,
-                fontSize: 14);
-    if ((icon ?? style?.value) != null) {
-      current = IconBox(
-          icon: icon ?? style?.value,
-          direction: currentOptions.direction,
-          spacing: currentOptions.spacing,
-          size: currentOptions.iconSize,
-          color: currentOptions.iconColor,
-          label: current);
+    final currentOptions = FlExtended().toastOptions.merge(options).copyWith(
+        duration: duration,
+        textStyle: textStyle,
+        animationDuration: animationDuration,
+        animationStyle: animationStyle,
+        decoration: decoration,
+        onToastTap: onToastTap,
+        color: color,
+        direction: direction,
+        builder: this.builder);
+
+    final builder = currentOptions.builder
+        ?.call(message, iconStyle?.icon, currentOptions.color);
+    Widget current = builder ?? const SizedBox();
+    if (builder == null) {
+      current = BText(message,
+          textAlign: TextAlign.center,
+          color: currentOptions.color,
+          maxLines: 5,
+          fontSize: 14);
+      if (iconStyle != null) {
+        current = IconBox(
+            icon: iconStyle!.icon,
+            direction: currentOptions.direction,
+            spacing: 10,
+            size: 18,
+            color: currentOptions.color,
+            label: current);
+      }
+      current = Padding(padding: const EdgeInsets.all(14), child: current);
     }
     current = Universal(
         safeTop: true,
@@ -51,12 +97,11 @@ class Toast extends StatelessWidget {
         safeLeft: true,
         onTap: currentOptions.onToastTap,
         decoration: currentOptions.decoration,
-        padding: currentOptions.padding,
         child: current);
-
     if (currentOptions.animationStyle != null) {
       current = FlAnimation(
           style: currentOptions.animationStyle!,
+          animationDuration: currentOptions.animationDuration,
           stayDuration: currentOptions.duration
               .subtract(kFlAnimationDuration, kFlAnimationDuration),
           child: current);
@@ -65,58 +110,93 @@ class Toast extends StatelessWidget {
   }
 }
 
+/// Toast类型
+/// 如果使用custom  请设置 [customIcon]
+enum ToastIconStyle {
+  /// success
+  success(FlExtendedIcons.success),
+
+  /// fail
+  fail(FlExtendedIcons.fail),
+
+  /// info
+  info(FlExtendedIcons.info),
+
+  /// warning
+  warning(FlExtendedIcons.warning),
+
+  /// smile
+  smile(FlExtendedIcons.smile);
+
+  const ToastIconStyle(this.icon);
+
+  final IconData icon;
+}
+
 /// Toast
 /// 关闭 closeToast();
 /// 添加 await Toast 关闭后继续执行之后的方法
-Future<ExtendedOverlayEntry?> showToast(String message,
-        {ToastStyle? style, IconData? icon, ToastOptions? options}) =>
-    Toast(message, options: options, icon: icon, style: style).show();
+Future<ExtendedOverlayEntry?> showToast(
+  String message, {
+  /// icon 样式 不传 仅显示文字
+  ToastIconStyle? iconStyle,
+
+  /// 动画样式
+  FlAnimationStyle? animationStyle,
+
+  /// 动画执行时间
+  Duration? animationDuration,
+
+  /// 显示文字样式
+  TextStyle? textStyle,
+
+  /// Toast显示时间
+  Duration? duration,
+
+  /// Toast 装饰器 会替换 [backgroundColor]
+  BoxDecoration? decoration,
+
+  /// Toast onTap
+  GestureTapCallback? onToastTap,
+
+  /// 重新 build  text icon
+  ToastOptionsBuilder? builder,
+
+  /// ToastOptions
+  ToastOptions? options,
+}) =>
+    Toast(
+      message,
+      animationStyle: animationStyle,
+      animationDuration: animationDuration,
+      textStyle: textStyle,
+      iconStyle: iconStyle,
+      builder: builder,
+      onToastTap: onToastTap,
+      decoration: decoration,
+      duration: duration,
+      options: options,
+    ).show();
 
 bool closeToast() => ExtendedOverlay().closeToast();
 
-/// Toast类型
-/// 如果使用custom  请设置 [customIcon]
-enum ToastStyle { success, fail, info, warning, smile }
-
-extension ExtensionToastStyle on ToastStyle {
-  Future<ExtendedOverlayEntry?> show(String message,
-          {IconData? icon, ToastOptions? options}) =>
-      Toast(message, options: options, icon: icon, style: this).show();
-
-  IconData get value {
-    switch (this) {
-      case ToastStyle.success:
-        return FlExtendedIcons.success;
-      case ToastStyle.fail:
-        return FlExtendedIcons.fail;
-      case ToastStyle.info:
-        return FlExtendedIcons.info;
-      case ToastStyle.warning:
-        return FlExtendedIcons.warning;
-      case ToastStyle.smile:
-        return FlExtendedIcons.smile;
-    }
-  }
-}
-
-typedef ToastOptionsBuildText = Widget Function(String text, Color color);
+typedef ToastOptionsBuilder = Widget Function(
+    String text, IconData? icon, Color? color);
 
 class ToastOptions extends BaseModalOptions {
   const ToastOptions({
-    this.iconColor = const Color(0xFFFFFFFF),
+    this.color = const Color(0xFFFFFFFF),
     this.decoration = const BoxDecoration(
         color: Color(0xEE000000),
         borderRadius: BorderRadius.all(Radius.circular(6))),
     this.onToastTap,
     this.textStyle,
     this.duration = const Duration(milliseconds: 1500),
+    this.animationDuration = const Duration(milliseconds: 300),
     this.animationStyle = FlAnimationStyle.verticalHunting,
-    this.iconSize = 18,
-    this.spacing = 10,
-    this.padding = const EdgeInsets.all(14),
     this.direction = Axis.horizontal,
-    this.buildText,
-    Color? modalColor,
+    this.builder,
+    super.backgroundColor,
     super.alignment = Alignment.center,
     super.onModalTap,
     super.gaussian = false,
@@ -126,19 +206,13 @@ class ToastOptions extends BaseModalOptions {
     super.filter,
     super.ignoring = false,
     super.absorbing = false,
-    super.canPop = true,
-    super.onPopInvoked,
-  }) : super(color: modalColor);
+  });
 
+  /// 动画样式
   final FlAnimationStyle? animationStyle;
 
-  final EdgeInsetsGeometry padding;
-
-  /// Toast 装饰器 会替换 [backgroundColor]
-  final BoxDecoration? decoration;
-
-  /// Toast onTap
-  final GestureTapCallback? onToastTap;
+  /// 动画执行时间
+  final Duration animationDuration;
 
   /// 显示文字样式
   final TextStyle? textStyle;
@@ -146,35 +220,31 @@ class ToastOptions extends BaseModalOptions {
   /// Toast显示时间
   final Duration duration;
 
-  /// icon
-  final Color iconColor;
+  /// Toast 装饰器 会替换 [backgroundColor]
+  final BoxDecoration? decoration;
 
-  /// icon size
-  final double iconSize;
+  /// Toast onTap
+  final GestureTapCallback? onToastTap;
 
-  /// Toast icon spacing
-  final double spacing;
+  /// 字体颜色和icon颜色
+  final Color color;
 
   /// Toast icon direction
   final Axis direction;
 
-  /// 重新 build Text
-  final ToastOptionsBuildText? buildText;
+  /// 重新 build  text icon
+  final ToastOptionsBuilder? builder;
 
   ToastOptions copyWith({
-    EdgeInsetsGeometry? padding,
     BoxDecoration? decoration,
     GestureTapCallback? onToastTap,
     TextStyle? textStyle,
     Duration? duration,
-    Color? iconColor,
-    double? iconSize,
-    double? spacing,
+    Color? color,
     Axis? direction,
-    ModalWindowsOptions? modalWindowsOptions,
-    ToastOptionsBuildText? buildText,
+    ToastOptionsBuilder? builder,
     GestureTapCallback? onModalTap,
-    Color? modalColor,
+    Color? backgroundColor,
     bool? addMaterial,
     ImageFilter? filter,
     bool? gaussian,
@@ -182,42 +252,42 @@ class ToastOptions extends BaseModalOptions {
     bool? ignoring,
     bool? absorbing,
     AlignmentGeometry? alignment,
+    FlAnimationStyle? animationStyle,
+    Duration? animationDuration,
   }) =>
       ToastOptions(
-          padding: padding ?? this.padding,
+          animationDuration: animationDuration ?? this.animationDuration,
+          animationStyle: animationStyle ?? this.animationStyle,
           decoration: decoration ?? this.decoration,
           onToastTap: onToastTap ?? this.onToastTap,
           textStyle: textStyle ?? this.textStyle,
           duration: duration ?? this.duration,
-          iconColor: iconColor ?? this.iconColor,
-          iconSize: iconSize ?? this.iconSize,
-          spacing: spacing ?? this.spacing,
-          buildText: buildText ?? this.buildText,
+          color: color ?? this.color,
+          builder: builder ?? this.builder,
           direction: direction ?? this.direction,
           ignoring: ignoring ?? this.ignoring,
           absorbing: ignoring ?? this.absorbing,
           alignment: alignment ?? this.alignment,
           onModalTap: onModalTap ?? this.onModalTap,
-          modalColor: modalColor ?? color,
+          backgroundColor: backgroundColor ?? this.backgroundColor,
           addMaterial: addMaterial ?? this.addMaterial,
           filter: filter ?? this.filter,
           gaussian: gaussian ?? this.gaussian,
           fuzzyDegree: fuzzyDegree ?? this.fuzzyDegree);
 
   ToastOptions merge([ToastOptions? options]) => copyWith(
-      padding: options?.padding,
+      animationDuration: options?.animationDuration,
+      animationStyle: options?.animationStyle,
       decoration: options?.decoration,
       onToastTap: options?.onToastTap,
       textStyle: options?.textStyle,
       duration: options?.duration,
-      iconColor: options?.iconColor,
-      iconSize: options?.iconSize,
-      spacing: options?.spacing,
-      buildText: options?.buildText,
+      color: options?.color,
+      builder: options?.builder,
       direction: options?.direction,
       alignment: options?.alignment,
       onModalTap: options?.onModalTap,
-      modalColor: options?.color,
+      backgroundColor: options?.color,
       ignoring: options?.ignoring,
       absorbing: options?.absorbing,
       addMaterial: options?.addMaterial,